Add ICEAgentPriv
