+++ lib/controlling_handler.ex
@@ -4,7 +4,7 @@ defmodule ExICE.ControllingHandler do
 
   require Logger
 
-  alias ExICE.{CandidatePair, Checklist, ICEAgent}
+  alias ExICE.{CandidatePair, Checklist, ICEAgentPriv}
   alias ExICE.Attribute.UseCandidate
 
   @impl true
@@ -12,22 +12,23 @@ defmodule ExICE.ControllingHandler do
     Logger.debug("""
     Received conn check request with use candidate attribute but
     we are the controlling side. Sending 400 bad request error response.
-    Pair: #{pair.id}. 
+    Pair: #{pair.id}.
     """)
 
-    ICEAgent.send_bad_request_error_response(pair, msg)
+    ICEAgentPriv.send_bad_request_error_response(pair, msg)
     state
   end
 
   @impl true
   def handle_conn_check_request(state, pair, msg, nil, key) do
-    ICEAgent.send_binding_success_response(pair, msg, key)
+    ICEAgentPriv.send_binding_success_response(pair, msg, key)
 
     # TODO use triggered check queue
     case Checklist.find_pair(state.checklist, pair) do
       nil ->
         Logger.debug("Adding new candidate pair: #{inspect(pair)}")
-        put_in(state, [:checklist, pair.id], pair)
+        checklist = Map.put(state.checklist, pair.id, pair)
+        %ICEAgentPriv{state | checklist: checklist}
 
       %CandidatePair{} = pair
       when state.selected_pair != nil and pair.discovered_pair_id == state.selected_pair.id ->
@@ -47,12 +48,12 @@ defmodule ExICE.ControllingHandler do
   @impl true
   def update_nominated_flag(%{eoc: true} = state, pair_id, true) do
     Logger.debug("Nomination succeeded. Selecting pair: #{inspect(pair_id)}")
-    state = ICEAgent.change_connection_state(:completed, state)
+    state = ICEAgentPriv.change_connection_state(:completed, state)
 
     pair = Map.fetch!(state.checklist, pair_id)
     pair = %CandidatePair{pair | nominate?: false, nominated?: true}
-
-    state = put_in(state, [:checklist, pair.id], pair)
+    checklist = Map.put(state.checklist, pair.id, pair)
+    state = %ICEAgentPriv{state | checklist: checklist}
 
     # the controlling agent could nominate only when eoc was set
     # and checklist finished
